Modify index.html and style.css
I modified the index.html page adding two divs to the header- one with a class name of right-items to group the notifcation bar and profile image to the right, and the other a class of left-items to group the links and the logo to the left. Next, I applied the follwoing css rule to both divs- display flex, align item center and gap of 16 px.To the header container in css i applied the following rules- display flex, justify content space between, and padding of 8px. Finally to ul i reset margin and padding to zero and then add display flex and a gap of 16px.